Factor out SRSValue, SRSSpec
