+++ dlog/commitment/src/srs.rs
@@ -112,3 +112,35 @@ impl<G: CommitmentCurve> SRS<G> where G::BaseField : PrimeField, G::ScalarField 
         Ok(SRS { g, h, endo_r, endo_q })
     }
 }
+
+pub enum SRSValue<'a, G : CommitmentCurve> {
+    Value(SRS<G>),
+    Ref(&'a SRS<G>)
+}
+
+impl<'a, G : CommitmentCurve> SRSValue<'a, G> {
+    pub fn get_ref(&self) -> & SRS<G> {
+        match self {
+            SRSValue::Value(x) => &x,
+            SRSValue::Ref(x) => x
+        }
+    }
+}
+
+pub enum SRSSpec <'a, G: CommitmentCurve>{
+    Use(&'a SRS<G>),
+    Generate(usize)
+}
+
+impl<'a, G: CommitmentCurve> SRSValue<'a, G> where G::BaseField : PrimeField, G::ScalarField : CommitmentField {
+    pub fn generate(size: usize) -> SRS<G> {
+        SRS::<G>::create(size)
+    }
+
+    pub fn create<'b>(spec : SRSSpec<'a, G>) -> SRSValue<'a, G>{
+        match spec {
+            SRSSpec::Use(x) => SRSValue::Ref(x),
+            SRSSpec::Generate(size) => SRSValue::Value(Self::generate(size))
+        }
+    }
+}
