+++ dlog/plonk-5-wires/src/index.rs
@@ -5,7 +5,7 @@ This source file implements Plonk Protocol Index primitive.
 *****************************************************************************************************************/
 
 use ff_fft::{DensePolynomial, Radix2EvaluationDomain as D};
-use commitment_dlog::{srs::SRS, CommitmentField, commitment::{CommitmentCurve, PolyComm}};
+use commitment_dlog::{srs::{SRSValue, SRSSpec}, CommitmentField, commitment::{CommitmentCurve, PolyComm}};
 use oracle::poseidon_5_wires::PlonkSpongeConstants;
 use oracle::poseidon::{ArithmeticSpongeParams, SpongeConstants};
 use plonk_5_wires_circuits::{constraints::{zk_w, ConstraintSystem}, wires::COLUMNS};
@@ -15,38 +15,6 @@ use algebra::PrimeField;
 
 type Fr<G> = <G as AffineCurve>::ScalarField;
 type Fq<G> = <G as AffineCurve>::BaseField;
-
-pub enum SRSValue<'a, G : CommitmentCurve> {
-    Value(SRS<G>),
-    Ref(&'a SRS<G>)
-}
-
-impl<'a, G : CommitmentCurve> SRSValue<'a, G> {
-    pub fn get_ref(&self) -> & SRS<G> {
-        match self {
-            SRSValue::Value(x) => &x,
-            SRSValue::Ref(x) => x
-        }
-    }
-}
-
-pub enum SRSSpec <'a, G: CommitmentCurve>{
-    Use(&'a SRS<G>),
-    Generate(usize)
-}
-
-impl<'a, G: CommitmentCurve> SRSValue<'a, G> where G::BaseField : PrimeField, G::ScalarField : CommitmentField {
-    pub fn generate(size: usize) -> SRS<G> {
-        SRS::<G>::create(size)
-    }
-
-    pub fn create<'b>(spec : SRSSpec<'a, G>) -> SRSValue<'a, G>{
-        match spec {
-            SRSSpec::Use(x) => SRSValue::Ref(x),
-            SRSSpec::Generate(size) => SRSValue::Value(Self::generate(size))
-        }
-    }
-}
 
 pub struct Index<'a, G: CommitmentCurve> where G::ScalarField : CommitmentField
 {
